requirements.txt
streamlit==1.31.0
Pillow==10.0.0
Werkzeug==2.0.1 